+++ packages/runtime/drqp_brain/launch/bringup.launch.py
@@ -57,7 +57,7 @@ def generate_launch_description():
                 name='load_drivers',
                 default_value='true',
                 choices=['true', 'false'],
-                description='Load drqp_a1_16_driver pose_setter',
+                description='Load drqp_control pose_setter',
             ),
             DeclareLaunchArgument(
                 name='servo_device',
@@ -65,7 +65,7 @@ def generate_launch_description():
                 description='Serial device for servos',
             ),
             Node(
-                package='drqp_a1_16_driver',
+                package='drqp_control',
                 executable='pose_setter',
                 output='screen',
                 parameters=[
