+++ packages/runtime/drqp_control/test/CMakeLists.txt
@@ -1,0 +1,10 @@
+
+find_package(ament_cmake_ros REQUIRED)
+find_package(launch_testing_ament_cmake REQUIRED)
+function(add_ros_isolated_launch_test path)
+  set(RUNNER "${ament_cmake_ros_DIR}/run_test_isolated.py")
+  add_launch_test("${path}" RUNNER "${RUNNER}" ${ARGN})
+endfunction()
+
+add_ros_isolated_launch_test(test_pose_reader.py)
+add_ros_isolated_launch_test(test_pose_setter.py)
