+++ packages/runtime/drqp_serial/include/drqp_serial/SerialFactory.h
@@ -1,4 +1,3 @@
-// Copyright (C) Pololu Corporation.
 // Copyright (c) 2017-2025 Anton Matosov
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
